Basic Matter implementation (#26)

+++ src/core/sizage.rs
@@ -1,13 +1,13 @@
 #[derive(Debug, PartialEq)]
 pub struct Sizage {
-    pub hs: u16,
-    pub ss: u16,
-    pub ls: u16,
-    pub fs: u16,
+    pub hs: u32,
+    pub ss: u32,
+    pub ls: u32,
+    pub fs: u32,
 }
 
 impl Sizage {
-    pub fn new(hs: u16, ss: u16, fs: u16, ls: u16) -> Sizage {
+    pub fn new(hs: u32, ss: u32, fs: u32, ls: u32) -> Sizage {
         Self { hs, ss, ls, fs }
     }
 }
